pyproject.toml
[project]
name = "serial-write-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "netifaces>=0.11.0",
    "psutil>=7.0.0",
    "pyserial>=3.5",
]
